Use route maximum speed.

+++ src/main/scala/logic/route.scala
@@ -20,7 +20,7 @@ class Road(
   val start: Town,
   val end: Town,
   val length: Double,
-  val maximum_speed: Int,
+  val maximum_speed: Double,
   val lanes: Int)
 extends Route {
   def accepts(vehicle: Vehicle): Boolean = {
@@ -61,7 +61,7 @@ class Rail(
   val start: Town,
   val end: Town,
   val length: Double,
-  val maximum_speed: Int,
+  val maximum_speed: Double,
   val tracks: Int,
   val electrified: Boolean)
 extends Route {
